requirements.txt
# Core framework
streamlit>=1.35.0

# LLM
litellm>=1.40.0

# Telegram
python-telegram-bot>=21.0

# Google APIs
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
gspread>=6.0.0

# Web tools
httpx>=0.27.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1

# Utilities
python-dotenv>=1.0.0
